feat(managed): Add transmit control to entity object

+++ managed/src/SwiftlyS2.Core/Modules/Schemas/Extensions/CBaseEntityImpl.cs
@@ -10,7 +10,7 @@ internal partial class CBaseEntityImpl : CBaseEntity
     {
         get
         {
-            return new CEntitySubclassVDataBaseImpl((nint)NativeSchema.GetVData(_Handle));
+            return new CEntitySubclassVDataBaseImpl(NativeSchema.GetVData(_Handle));
         }
     }
